+++ server/src/models/event-post-settings.model.ts
