+++ server/src/models/interfaces/auth-info.ts
