+++ server/src/models/interfaces/discord-channel.ts
