+++ server/src/models/interfaces/discord-embed.ts
