+++ server/src/models/interfaces/discord-member.ts
