+++ server/src/models/interfaces/discord-message-details.ts
@@ -1,6 +1,7 @@
-import DiscordEmbed from './discordembed.interface';
+import DiscordEmbed from './discord-embed';
 
 export interface DiscordMessageDetails {
   id: string;
   embeds: DiscordEmbed[];
 }
+
