+++ server/src/models/interfaces/discord-message.ts
