+++ server/src/models/interfaces/discord-role.ts
