+++ server/src/models/interfaces/formatted-log-entry.ts
