+++ server/src/models/interfaces/gw2-log-entry.ts
