+++ server/src/models/interfaces/gw2-member.ts
