+++ server/src/models/interfaces/gw2-rank.ts
