+++ server/src/services/gw2/guild-api.ts
@@ -1,9 +1,9 @@
 import { Service } from 'typedi';
 import { config } from '../../config';
-import GW2LogEntry from '../../models/interfaces/gw2logentry.interface';
-import GW2Member from '../../models/interfaces/gw2member.interface';
-import GW2Rank from '../../models/interfaces/gw2rank.interface';
-import { GW2Api } from './api.gw2.service';
+import GW2LogEntry from '../../models/interfaces/gw2-log-entry';
+import GW2Member from '../../models/interfaces/gw2-member';
+import GW2Rank from '../../models/interfaces/gw2-rank';
+import { GW2Api } from './gw2-api';
 
 @Service()
 export class GW2GuildApi {
@@ -29,3 +29,4 @@ export class GW2GuildApi {
     return await this.gw2Api.get(`guild/${this.guildId}/${endpoint}`);
   }
 }
+
