+++ server/src/services/repositories/event-repository.ts
@@ -2,7 +2,7 @@ import { Service } from 'typedi';
 import { Repository } from 'typeorm';
 import { InjectRepository } from 'typeorm-typedi-extensions';
 import { Event } from '../../models/event.model';
-import { BaseRepository } from './base.repository';
+import { BaseRepository } from './base-repository';
 
 @Service()
 export class EventRepository extends BaseRepository<Event> {
@@ -14,3 +14,4 @@ export class EventRepository extends BaseRepository<Event> {
     return await this.repo.find({ day });
   }
 }
+
